Accidentally exposed API Keys
These were stored in a .env file. In the old .gitignore, .env was
ignored. In the new file, it isn't. This resulted in accidentally
pushing the Slack API keys to the repo. Whoops.

Fortunately, Slack was on top of this and killed the API tokens right
away. Anyway, this commit removes the .env file from the repo and
re-adds that file to .gitignore.